+++ spec/app/models/sandwich.rb
@@ -1,0 +1,4 @@
+class Sandwich
+  include Mongoid::Document
+  has_and_belongs_to_many :meats
+end