+++ spec/mongoid/relations/referenced/many_to_many_spec.rb
@@ -1696,6 +1696,25 @@ describe Mongoid::Relations::Referenced::ManyToMany do
       Person.create
     end
 
+    context "when nothing exists on the relation" do
+
+      context "when the document is destroyed" do
+
+        before do
+          Meat.create!
+        end
+
+        let!(:sandwich) do
+          Sandwich.create!
+        end
+
+        it "returns zero" do
+          sandwich.destroy
+          sandwich.meats.count.should eq(0)
+        end
+      end
+    end
+
     context "when documents have been persisted" do
 
       let!(:preference) do